extract following box
