+++ components/FollowingProfileBox/index.js
@@ -1,0 +1,3 @@
+import FollowingProfileBox from './FollowingProfileBox';
+
+export default FollowingProfileBox;
